requirements.txt
Click==7.0
cycler==0.10.0
dnspython==1.16.0
eventlet==0.24.1
ezdxf==0.9
Flask==1.0.2
Flask-MQTT==1.0.5
Flask-SocketIO==3.3.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
monotonic==1.5
numpy==1.16.3
paho-mqtt==1.4.0
pkg-resources==0.0.0
pygame==1.9.6
pyparsing==2.4.0
python-dateutil==2.8.0
python-engineio==3.5.1
python-socketio==3.1.2
six==1.12.0
typing==3.6.6
Werkzeug==0.15.2
